requirements.txt
fastapi
sqlmodel
uvicorn
python-dotenv
google-generativeai
pydantic
python-dateutil
psycopg2-binary
